+++ src/packages/excalidraw/index.tsx
@@ -195,6 +195,7 @@ export {
   exportToBlob,
   exportToSvg,
   serializeAsJSON,
+  serializeLibraryAsJSON,
   loadLibraryFromBlob,
   loadFromBlob,
   getFreeDrawSvgPath,