+++ src/packages/utils.ts
@@ -139,7 +139,7 @@ export const exportToSvg = async ({
   );
 };
 
-export { serializeAsJSON } from "../data/json";
+export { serializeAsJSON, serializeLibraryAsJSON } from "../data/json";
 export { loadFromBlob, loadLibraryFromBlob } from "../data/blob";
 export { getFreeDrawSvgPath } from "../renderer/renderElement";
 export { getCommonBoundingBox } from "../element/bounds"; //zsviczian