+++ src/tests/data/restore.test.ts
@@ -314,9 +314,7 @@ describe("restoreAppState", () => {
     expect(restoredAppState.activeTool).toEqual(
       stubImportedAppState.activeTool,
     );
-    expect(restoredAppState.cursorButton).toBe(
-      stubImportedAppState.cursorButton,
-    );
+    expect(restoredAppState.cursorButton).toBe("up");
     expect(restoredAppState.name).toBe(stubImportedAppState.name);
   });
 
@@ -348,9 +346,7 @@ describe("restoreAppState", () => {
       stubImportedAppState,
       null,
     );
-    expect(restoredAppState.cursorButton).toBe(
-      stubImportedAppState.cursorButton,
-    );
+    expect(restoredAppState.cursorButton).toBe("up");
     expect(restoredAppState.name).toBe(stubImportedAppState.name);
   });
 
@@ -503,9 +499,7 @@ describe("restore", () => {
     importedDataState.appState = stubImportedAppState;
 
     const restoredData = restore.restore(importedDataState, null, null);
-    expect(restoredData.appState.cursorButton).toBe(
-      stubImportedAppState.cursorButton,
-    );
+    expect(restoredData.appState.cursorButton).toBe("up");
     expect(restoredData.appState.name).toBe(stubImportedAppState.name);
   });