+++ src/tests/helpers/api.ts
@@ -15,6 +15,7 @@ import path from "path";
 import { getMimeType } from "../../data/blob";
 import { newFreeDrawElement } from "../../element/newElement";
 import { Point } from "../../types";
+import { getSelectedElements } from "../../scene/selection";
 
 const readFile = util.promisify(fs.readFile);
 
@@ -30,10 +31,10 @@ export class API {
     });
   };
 
-  static getSelectedElements = (): ExcalidrawElement[] => {
-    return h.elements.filter(
-      (element) => h.state.selectedElementIds[element.id],
-    );
+  static getSelectedElements = (
+    includeBoundTextElement: boolean = false,
+  ): ExcalidrawElement[] => {
+    return getSelectedElements(h.elements, h.state, includeBoundTextElement);
   };
 
   static getSelectedElement = (): ExcalidrawElement => {
@@ -101,6 +102,7 @@ export class API {
       ? ExcalidrawTextElement["containerId"]
       : never;
     points?: T extends "arrow" | "line" ? readonly Point[] : never;
+    locked?: boolean;
   }): T extends "arrow" | "line"
     ? ExcalidrawLinearElement
     : T extends "freedraw"
@@ -126,6 +128,7 @@ export class API {
       roughness: rest.roughness ?? appState.currentItemRoughness,
       opacity: rest.opacity ?? appState.currentItemOpacity,
       boundElements: rest.boundElements ?? null,
+      locked: rest.locked ?? false,
     };
     switch (type) {
       case "rectangle":